port to macOS (by Cursor + Gemini 2.5 Pro)

+++ magiskbase/stream.cpp
@@ -1,19 +1,20 @@
 #include <unistd.h>
 #include <cstddef>
+#include <sys/uio.h>
 
 #include <base.hpp>
 #include <stream.hpp>
 
 using namespace std;
 
-static int strm_read(void *v, char *buf, size_t len) {
+static int strm_read(void *v, char *buf, int len) {
     auto strm = static_cast<stream *>(v);
-    return strm->read(buf, len);
+    return strm->read(buf, static_cast<size_t>(len));
 }
 
-static int strm_write(void *v, const char *buf, size_t len) {
+static int strm_write(void *v, const char *buf, int len) {
     auto strm = static_cast<stream *>(v);
-    if (!strm->write(buf, len))
+    if (!strm->write(buf, static_cast<size_t>(len)))
         return -1;
     return len;
 }
